Fixed errors and add feign client and open api

+++ delivery-service/src/main/resources/delivery_schema.sql
@@ -6,7 +6,7 @@ GRANT USAGE ON SCHEMA delivery_service TO ms_user;
 --------------------
 
 create table delivery_service.deliveries (
-                                      id bigint primary key,
+                                      id serial primary key,
                                       order_id bigint not null,
                                       status varchar(16) not null,
                                       street varchar(255) not null,
